Move predicate to a mod
